chore: moved packet for shore
